+++ src/main/java/org/example/tackit/domain/QnA_board/QnA_comment/repository/QnACommentRepository.java
@@ -1,0 +1,14 @@
+package org.example.tackit.domain.QnA_board.QnA_comment.repository;
+
+import org.example.tackit.domain.entity.QnAComment;
+import org.example.tackit.domain.entity.QnAPost;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface QnACommentRepository extends JpaRepository<QnAComment, Long> {
+    List<QnAComment> findByQnAPost(QnAPost post);
+}
+
