+++ src/main/java/org/example/tackit/domain/QnA_board/QnA_post/dto/request/QnAPostRequestDto.java
@@ -1,0 +1,14 @@
+package org.example.tackit.domain.QnA_board.QnA_post.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class QnAPostRequestDto {
+    private String title;
+    private String content;
+    private String tag;
+}
