+++ src/main/java/org/example/tackit/domain/QnA_board/QnA_tag/repository/QnATagRepository.java
@@ -1,0 +1,7 @@
+package org.example.tackit.domain.QnA_board.QnA_tag.repository;
+
+import org.example.tackit.domain.entity.QnATag;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface QnATagRepository extends JpaRepository<QnATag, Long> {
+}
