--- src/main/java/org/example/tackit/domain/QnA_post/dto/request/UpdateQnARequestDto.java
@@ -1,14 +1,0 @@
-package org.example.tackit.domain.QnA_post.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class UpdateQnARequestDto {
-    private String title;
-    private String content;
-    private String tag;
-}
