--- src/main/java/org/example/tackit/domain/QnA_post/repository/QnAMemberRepository.java
@@ -1,14 +1,0 @@
-package org.example.tackit.domain.QnA_post.repository;
-
-import org.example.tackit.domain.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface QnAMemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByNickname(String nickname);
-    Optional<Member> findByEmail(String email);
-}
-
