+++ src/main/java/org/example/tackit/domain/entity/QnATag.java
@@ -1,0 +1,23 @@
+package org.example.tackit.domain.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Table(name="qna_tag", uniqueConstraints = @UniqueConstraint(columnNames = "tagName") )
+public class QnATag {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column (nullable = false)
+    private String tagName;
+
+}
